requirements.txt
python-dotenv~=1.1.0
langchain-google-genai~=2.1.5
qdrant-client~=1.14.2
langchain-community~=0.3.25
langchain-text-splitters~=0.3.8
pandas~=2.2.3
docx~=0.2.4